Split event dispatcher class in gRPC service and generic dispatcher
