+++ axonserver/src/main/java/io/axoniq/axonserver/message/event/EventDispatcher.java
@@ -10,25 +10,17 @@
 package io.axoniq.axonserver.message.event;
 
 import io.axoniq.axonserver.applicationevents.TopologyEvents;
-import io.axoniq.axonserver.config.AuthenticationProvider;
 import io.axoniq.axonserver.exception.ExceptionUtils;
 import io.axoniq.axonserver.exception.MessagingPlatformException;
-import io.axoniq.axonserver.grpc.AxonServerClientService;
-import io.axoniq.axonserver.grpc.ContextProvider;
 import io.axoniq.axonserver.grpc.GrpcExceptionBuilder;
 import io.axoniq.axonserver.grpc.GrpcFlowControlExecutorProvider;
 import io.axoniq.axonserver.grpc.event.Confirmation;
 import io.axoniq.axonserver.grpc.event.Event;
-import io.axoniq.axonserver.grpc.event.EventStoreGrpc;
 import io.axoniq.axonserver.grpc.event.GetAggregateEventsRequest;
 import io.axoniq.axonserver.grpc.event.GetAggregateSnapshotsRequest;
 import io.axoniq.axonserver.grpc.event.GetEventsRequest;
-import io.axoniq.axonserver.grpc.event.GetFirstTokenRequest;
-import io.axoniq.axonserver.grpc.event.GetLastTokenRequest;
-import io.axoniq.axonserver.grpc.event.GetTokenAtRequest;
 import io.axoniq.axonserver.grpc.event.QueryEventsRequest;
 import io.axoniq.axonserver.grpc.event.QueryEventsResponse;
-import io.axoniq.axonserver.grpc.event.ReadHighestSequenceNrRequest;
 import io.axoniq.axonserver.grpc.event.ReadHighestSequenceNrResponse;
 import io.axoniq.axonserver.grpc.event.TrackingToken;
 import io.axoniq.axonserver.localstorage.SerializedEvent;
@@ -39,8 +31,6 @@ import io.axoniq.axonserver.topology.EventStoreLocator;
 import io.axoniq.axonserver.util.StreamObserverUtils;
 import io.axoniq.flowcontrol.OutgoingStream;
 import io.axoniq.flowcontrol.producer.grpc.FlowControlledOutgoingStream;
-import io.grpc.MethodDescriptor;
-import io.grpc.protobuf.ProtoUtils;
 import io.grpc.stub.CallStreamObserver;
 import io.grpc.stub.StreamObserver;
 import io.micrometer.core.instrument.Tags;
@@ -70,43 +60,20 @@ import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
 
 import static io.axoniq.axonserver.exception.ErrorCode.NO_EVENTSTORE;
-import static io.grpc.stub.ServerCalls.*;
 
 /**
  * @author Marc Gathier
  * @author Sara Pellegrini
  * @author Stefan Dragisic
  */
-@Component("EventDispatcher")
-public class EventDispatcher implements AxonServerClientService {
+@Component
+public class EventDispatcher  {
 
-    public static final MethodDescriptor<GetEventsRequest, InputStream> METHOD_LIST_EVENTS =
-            EventStoreGrpc.getListEventsMethod().toBuilder(
-                    ProtoUtils.marshaller(GetEventsRequest.getDefaultInstance()),
-                    InputStreamMarshaller.inputStreamMarshaller())
-                    .build();
-    public static final MethodDescriptor<GetAggregateEventsRequest, SerializedEvent> METHOD_LIST_AGGREGATE_EVENTS =
-            EventStoreGrpc.getListAggregateEventsMethod().toBuilder(
-                    ProtoUtils.marshaller(GetAggregateEventsRequest.getDefaultInstance()),
-                    SerializedEventMarshaller.serializedEventMarshaller())
-                    .build();
-    public static final MethodDescriptor<GetAggregateSnapshotsRequest, SerializedEvent> METHOD_LIST_AGGREGATE_SNAPSHOTS =
-            EventStoreGrpc.getListAggregateSnapshotsMethod().toBuilder(
-                    ProtoUtils.marshaller(GetAggregateSnapshotsRequest.getDefaultInstance()),
-                    SerializedEventMarshaller.serializedEventMarshaller())
-                    .build();
-    public static final MethodDescriptor<InputStream, Confirmation> METHOD_APPEND_EVENT =
-            EventStoreGrpc.getAppendEventMethod().toBuilder(
-                    InputStreamMarshaller.inputStreamMarshaller(),
-                    ProtoUtils.marshaller(Confirmation.getDefaultInstance()))
-                    .build();
     static final String ERROR_ON_CONNECTION_FROM_EVENT_STORE = "{}:  Error on connection from event store: {}";
     private static final String NO_EVENT_STORE_CONFIGURED = "No event store available for: ";
     private final Logger logger = LoggerFactory.getLogger(EventDispatcher.class);
     private final EventStoreLocator eventStoreLocator;
-    private final AuthenticationProvider authenticationProvider;
     private final MeterFactory meterFactory;
-    private final ContextProvider contextProvider;
     private final Map<ClientStreamIdentification, List<EventTrackerInfo>> trackingEventProcessors = new ConcurrentHashMap<>();
     private final Map<String, MeterFactory.RateMeter> eventsCounter = new ConcurrentHashMap<>();
     private final Map<String, MeterFactory.RateMeter> snapshotCounter = new ConcurrentHashMap<>();
@@ -117,28 +84,18 @@ public class EventDispatcher implements AxonServerClientService {
     private final int aggregateEventsPrefetch;
 
     public EventDispatcher(EventStoreLocator eventStoreLocator,
-                           ContextProvider contextProvider,
-                           AuthenticationProvider authenticationProvider,
                            MeterFactory meterFactory,
                            GrpcFlowControlExecutorProvider grpcFlowControlExecutorProvider,
                            @Value("${axoniq.axonserver.event.aggregate.retry.attempts:3}") int maxRetryAttempts,
                            @Value("${axoniq.axonserver.event.aggregate.retry.delay:100}") long retryDelayMillis,
                            @Value("${axoniq.axonserver.event.aggregate.prefetch:5}") int aggregateEventsPrefetch) {
-        this.contextProvider = contextProvider;
         this.eventStoreLocator = eventStoreLocator;
-        this.authenticationProvider = authenticationProvider;
         this.meterFactory = meterFactory;
         this.grpcFlowControlExecutorProvider = grpcFlowControlExecutorProvider;
         retrySpec = Retry.backoff(maxRetryAttempts, Duration.ofMillis(retryDelayMillis));
         this.aggregateEventsPrefetch = aggregateEventsPrefetch;
     }
 
-
-    public StreamObserver<InputStream> appendEvent(StreamObserver<Confirmation> responseObserver) {
-        CallStreamObserver<Confirmation> callStreamObserver = (CallStreamObserver<Confirmation>) responseObserver;
-        return appendEvent(contextProvider.getContext(), authenticationProvider.get(),
-                new ForwardingStreamObserver<>(logger, "appendEvent", callStreamObserver));
-    }
 
     public StreamObserver<InputStream> appendEvent(String context, Authentication authentication,
                                                    StreamObserver<Confirmation> responseObserver) {
@@ -203,14 +160,6 @@ public class EventDispatcher implements AxonServerClientService {
     }
 
 
-    public void appendSnapshot(Event event, StreamObserver<Confirmation> streamObserver) {
-        CallStreamObserver<Confirmation> callStreamObserver = (CallStreamObserver<Confirmation>) streamObserver;
-        appendSnapshot(contextProvider.getContext(),
-                authenticationProvider.get(),
-                event,
-                new ForwardingStreamObserver<>(logger, "appendSnapshot", callStreamObserver));
-    }
-
     public void appendSnapshot(String context, Authentication authentication, Event snapshot,
                                StreamObserver<Confirmation> responseObserver) {
         checkConnection(context, responseObserver).ifPresent(eventStore -> {
@@ -235,18 +184,6 @@ public class EventDispatcher implements AxonServerClientService {
                 responseObserver.onError(ex);
             }
         });
-    }
-
-
-    public void listAggregateEvents(GetAggregateEventsRequest request,
-                                    StreamObserver<SerializedEvent> responseObserver) {
-        CallStreamObserver<SerializedEvent> streamObserver = (CallStreamObserver<SerializedEvent>) responseObserver;
-        String context = contextProvider.getContext();
-        CallStreamObserver<SerializedEvent> validateStreamObserver =
-                new SequenceValidationStreamObserver(streamObserver, sequenceValidationStrategy, context);
-        listAggregateEvents(context, authenticationProvider.get(),
-                            request,
-                            new ForwardingStreamObserver<>(logger, "listAggregateEvents", validateStreamObserver));
     }
 
 
@@ -314,10 +251,6 @@ public class EventDispatcher implements AxonServerClientService {
         return eventStore.aggregateEvents(context, principal, request);
     }
 
-    public StreamObserver<GetEventsRequest> listEvents(StreamObserver<InputStream> responseObserver) {
-        return listEvents(contextProvider.getContext(), authenticationProvider.get(), responseObserver);
-    }
-
     public StreamObserver<GetEventsRequest> listEvents(String context, Authentication principal,
                                                        StreamObserver<InputStream> responseObserver) {
         return new GetEventsRequestStreamObserver(responseObserver, context, principal);
@@ -364,49 +297,9 @@ public class EventDispatcher implements AxonServerClientService {
     }
 
 
-    @Override
-    public final io.grpc.ServerServiceDefinition bindService() {
-        return io.grpc.ServerServiceDefinition.builder(EventStoreGrpc.SERVICE_NAME)
-                .addMethod(
-                        METHOD_APPEND_EVENT,
-                        asyncClientStreamingCall(this::appendEvent))
-                .addMethod(
-                        EventStoreGrpc.getAppendSnapshotMethod(),
-                        asyncUnaryCall(this::appendSnapshot))
-                .addMethod(
-                        METHOD_LIST_AGGREGATE_EVENTS,
-                        asyncServerStreamingCall(this::listAggregateEvents))
-                .addMethod(
-                        METHOD_LIST_AGGREGATE_SNAPSHOTS,
-                        asyncServerStreamingCall(this::listAggregateSnapshots))
-                .addMethod(
-                        METHOD_LIST_EVENTS,
-                        asyncBidiStreamingCall(this::listEvents))
-                .addMethod(
-                        EventStoreGrpc.getReadHighestSequenceNrMethod(),
-                        asyncUnaryCall(this::readHighestSequenceNr))
-                .addMethod(
-                        EventStoreGrpc.getGetFirstTokenMethod(),
-                        asyncUnaryCall(this::getFirstToken))
-                .addMethod(
-                        EventStoreGrpc.getGetLastTokenMethod(),
-                        asyncUnaryCall(this::getLastToken))
-                .addMethod(
-                        EventStoreGrpc.getGetTokenAtMethod(),
-                        asyncUnaryCall(this::getTokenAt))
-                .addMethod(
-                        EventStoreGrpc.getQueryEventsMethod(),
-                        asyncBidiStreamingCall(this::queryEvents))
-                .build();
-    }
-
-    public void getFirstToken(GetFirstTokenRequest request, StreamObserver<TrackingToken> streamObserver) {
-        CallStreamObserver<TrackingToken> callStreamObserver = (CallStreamObserver<TrackingToken>) streamObserver;
-        ForwardingStreamObserver<TrackingToken> responseObserver = new ForwardingStreamObserver<>(logger,
-                "getFirstToken",
-                callStreamObserver);
-        checkConnection(contextProvider.getContext(), responseObserver)
-                .ifPresent(client -> client.firstEventToken(contextProvider.getContext())
+    public void getFirstToken(String context, StreamObserver<TrackingToken> responseObserver) {
+        checkConnection(context, responseObserver)
+                .ifPresent(client -> client.firstEventToken(context)
                                            .map(token -> TrackingToken.newBuilder().setToken(token).build())
                                            .subscribe(responseObserver::onNext,
                                                       responseObserver::onError,
@@ -423,13 +316,9 @@ public class EventDispatcher implements AxonServerClientService {
         return Optional.of(eventStore);
     }
 
-    public void getLastToken(GetLastTokenRequest request, StreamObserver<TrackingToken> streamObserver) {
-        CallStreamObserver<TrackingToken> callStreamObserver = (CallStreamObserver<TrackingToken>) streamObserver;
-        ForwardingStreamObserver<TrackingToken> responseObserver = new ForwardingStreamObserver<>(logger,
-                "getLastToken",
-                callStreamObserver);
-        checkConnection(contextProvider.getContext(), responseObserver)
-                .ifPresent(client -> client.lastEventToken(contextProvider.getContext())
+    public void getLastToken(String context, StreamObserver<TrackingToken> responseObserver) {
+        checkConnection(context, responseObserver)
+                .ifPresent(client -> client.lastEventToken(context)
                                            .map(token -> TrackingToken.newBuilder().setToken(token).build())
                                            .subscribe(responseObserver::onNext,
                                                       responseObserver::onError,
@@ -437,28 +326,21 @@ public class EventDispatcher implements AxonServerClientService {
                 );
     }
 
-    public void getTokenAt(GetTokenAtRequest request, StreamObserver<TrackingToken> streamObserver) {
-        CallStreamObserver<TrackingToken> callStreamObserver = (CallStreamObserver<TrackingToken>) streamObserver;
-        ForwardingStreamObserver<TrackingToken> responseObserver = new ForwardingStreamObserver<>(logger,
-                "getTokenAt",
-                callStreamObserver);
-        checkConnection(contextProvider.getContext(), responseObserver)
-                .ifPresent(client -> client.eventTokenAt(contextProvider.getContext(),
-                                                         Instant.ofEpochMilli(request.getInstant()))
+    public void getTokenAt(String context, long instant, StreamObserver<TrackingToken> responseObserver) {
+        checkConnection(context, responseObserver)
+                .ifPresent(client -> client.eventTokenAt(context,
+                                                         Instant.ofEpochMilli(instant))
                                            .map(token -> TrackingToken.newBuilder().setToken(token).build())
                                            .subscribe(responseObserver::onNext,
                                                       responseObserver::onError,
                                                       responseObserver::onCompleted));
     }
 
-    public void readHighestSequenceNr(ReadHighestSequenceNrRequest request,
-                                      StreamObserver<ReadHighestSequenceNrResponse> streamObserver) {
-        CallStreamObserver<ReadHighestSequenceNrResponse> callStreamObserver = (CallStreamObserver<ReadHighestSequenceNrResponse>) streamObserver;
-        ForwardingStreamObserver<ReadHighestSequenceNrResponse> responseObserver =
-                new ForwardingStreamObserver<>(logger, "readHighestSequenceNr", callStreamObserver);
-        checkConnection(contextProvider.getContext(), responseObserver)
+    public void readHighestSequenceNr(String context, String aggregateId,
+                                      StreamObserver<ReadHighestSequenceNrResponse> responseObserver) {
+        checkConnection(context, responseObserver)
                 .ifPresent(client -> client
-                        .highestSequenceNumber(contextProvider.getContext(), request.getAggregateId())
+                        .highestSequenceNumber(context, aggregateId)
                         .map(l -> ReadHighestSequenceNrResponse.newBuilder().setToSequenceNr(l).build())
                         .subscribe(responseObserver::onNext,
                                    responseObserver::onError,
@@ -466,12 +348,9 @@ public class EventDispatcher implements AxonServerClientService {
                 );
     }
 
-    public StreamObserver<QueryEventsRequest> queryEvents(StreamObserver<QueryEventsResponse> streamObserver) {
-        CallStreamObserver<QueryEventsResponse> callStreamObserver = (CallStreamObserver<QueryEventsResponse>) streamObserver;
-        String context = contextProvider.getContext();
-        Authentication authentication = authenticationProvider.get();
-        ForwardingStreamObserver<QueryEventsResponse> responseObserver =
-                new ForwardingStreamObserver<>(logger, "queryEvents", callStreamObserver);
+    public StreamObserver<QueryEventsRequest> queryEvents(String context,
+                                                          Authentication authentication,
+                                                          StreamObserver<QueryEventsResponse> responseObserver) {
         return new StreamObserver<QueryEventsRequest>() {
 
             private final AtomicReference<Sinks.Many<QueryEventsRequest>> requestSinkRef = new AtomicReference<>();
@@ -541,12 +420,6 @@ public class EventDispatcher implements AxonServerClientService {
             }
         });
     }
-
-    public void listAggregateSnapshots(GetAggregateSnapshotsRequest request,
-                                       StreamObserver<SerializedEvent> responseObserver) {
-        listAggregateSnapshots(contextProvider.getContext(), authenticationProvider.get(), request, responseObserver);
-    }
-
 
     public MeterFactory.RateMeter eventRate(String context) {
         return eventsCounter(context, eventsCounter, BaseMetricName.AXON_EVENTS);
